fix(navigation): make the mobile menu a modal and split the project combobox
The mobile menu overlay now carries dialog semantics: it takes focus on
open, closes on Escape, and returns focus to the menu button. Its bar
button prop is named for what it does instead of promising a toggle.

The project combobox moves its item building and its popup into their
own modules, so each file stays inside the size limits, and its search
input carries an accessible name of its own.

Also renames the sidebar fullWidth prop to isFullWidth, and adds
coverage for the single-select scope picker search.
